pyproject.toml
[tool.poetry]
name = "insight-py"
version = "0.1.0"
description = "Python port of the Insight MCP server"
authors = ["Atomik"]
license = "MIT"
packages = [{include = "insight", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
mcp = {extras = ["cli"], version = "^1.2.0"}
langchain-community = "^0.3.15"
langchain-openai = "^0.3.2"
langchain-anthropic = "^0.3.4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
insight-py = "insight_py.__main__:main"
